Don't reload seed on file load if not changing seeds
This change means that the randomizer itself is never cleared from memory, and thus any checks for this have been changed to use the new randoIsEnabled function.

Fixed a bug in loadCustomText that would cause an incorrect index to be used if some languages were excluded from the seed GCI.

Removed instances of setting cleared pointers to nullptr when unloaded a seed, as they don't actually get used once the seed is unloaded.