Add next changes
Added -fvisibility=hidden and --gc-keep-exported to the Makefile in place of _sector_anchor and _sector_anchor_var

Changed KEEP_FUNC and KEEP_VAR to use the visibility attritute to tag along with the Makefile changes.

The visibility attribute will not work correctly if placed after a type for a pointer, so all instances of the defines were placed in front of the variables/functions.

Made a few adjustments to the seedlist functions/files.

Added applicable modifyItemData code to the boot rel.