Fix console display issues with seeds
"Please select a seed" is now displayed properly when the player must choose a seed to use.

Some values were casted to ints when being printed to the console to avoid leading 0s on them.

"Press X/Y to select a seed" is no longer displayed at boot since this is redundant.