Add most languages/code for item wheel menu
The Japanese texts for Snowpeak Ruins and Faron Woods does not display properly, so these texts will be changed in the future.

The Japanese text for pressing buttons to toggle the menu needs to be adjusted to be on multiple lines.

Spanish is still missing, so English text is being used in the mean time.

The memory allocated for m_MsgTableInfo was changed to align to the size of the offsets, as anything higher is unnecessary.

Changed uses of /x to \x in frenchMessages.cpp.

Adjusted where the clang formatting should restart in the messages cpp files.

Added the necessary symbols for building the Japanese version.