Add current work for custom messages that can take variadic params
Currently only the English text has been adjusted for this.

The color of the areas for the dungeon items will need to be manually adjusted just before the string has been created. This can be done by setting up a global variable that contains the exact offset into the string in which the color value is, and then replacing it based on the current area. This replacement would be performed on the original string, in which the cache would need to be cleared via DCFlushRange after being modified.