Fix being able to lose Poe items
This causes defeated Poes (excluding Arbiter Poes) to immediately despawn, and the item they would give is added to the queue to give to the player when they are free.

Also adjusted _00_handle_poeItem to not increment the Poe count if it is already at 60 or greater.

Also increased the amount of bytes used for queued items to 8.