Clear reserved bytes upon reaching title screen
This is because the reserved bytes that the queue uses to store the items to give are not cleared upon starting a new file, which means that the player could soft reset during the process of being given item(s), and then start a new file to be given those items on that new file.

Also moved resetting giveItemToPlayer to QUEUE_EMPTY to resetQueueOnFileSelectScreen as well, since it makes more sense for it to be reset there alongside the reserved bytes being cleared.